[CHORE] 판매량 상위 상품  도메인 및 dto 필드 수정

+++ src/main/java/kr/hhplus/be/server/interfaces/web/bestseller/dto/response/BestSellerResponse.java
@@ -10,7 +10,7 @@ public class BestSellerResponse {
     private Long id;
     private String name;
     private Long price;
-    private Integer stock;
+    private Long stock;
     private Long rank;
 
     public static BestSellerResponse from(BestSeller bestSeller) {
